+++ commands/fun/coinflip.js
@@ -1,6 +1,6 @@
 const Discord = require('discord.js');
 const embed = new Discord.MessageEmbed();
-const conf = require("../../configs/global.json");
+const conf = require("../../utils/global.json");
 
 module.exports.run = async (bot, message, args) => {
   // thanks to Will for this code.
@@ -35,7 +35,9 @@ module.exports.run = async (bot, message, args) => {
   }
 
   embed.setColor('RANDOM');
-  embed.setTitle(start[srtat] + conf.md + result + conf.md);
+  embed.setTitle('Coinflip');
+  embed.setDescription(start[srtat] + conf.md + result + conf.md);
+  embed.setTimestamp();
   
   message.reply(embed);
 }