+++ commands/fun/dankmeme.js
@@ -9,13 +9,13 @@ module.exports.run = async (bot, message, args) => {
 		embed.setColor('RANDOM');
 		embed.setTitle(exports.help.name + " help.");
 		embed.addField("layout:", config.prefix + exports.help.name + "", false);
-		embed.addField("purpose:", "shows a random fun fact from reddit.", false);
+		embed.addField("purpose:", "shows a dank meme from reddit.", false);
 		message.channel.send(embed);
 		return;
 	
 	  }
 
-	got('https://www.reddit.com/r/funfacts/random/.json')
+	got('https://www.reddit.com/r/dankmemes/random/.json')
 		.then(response => {
 			const [list] = JSON.parse(response.body);
 			const [post] = list.data.children;
@@ -40,6 +40,6 @@ module.exports.run = async (bot, message, args) => {
 };
 
 module.exports.help = {
-	name: "fact", // * Dwight: Fact, this is js.
-	aliases: ['interesting']
+	name: "dankmeme",
+	aliases: ['dank']
 };