+++ commands/fun/time.js
@@ -1,5 +1,5 @@
 const Discord = require('discord.js');
-const conf = require('../../configs/global.json');
+const conf = require('../../utils/global.json');
 const embed = new Discord.MessageEmbed();
 var date = new Date();
 
@@ -41,7 +41,9 @@ module.exports.run = async (bot, message, args) => {
 
 
     embed.setColor('RANDOM');
-    embed.setTitle(start[srtat] + conf.md + time + " UTC" + tz + conf.md);
+    embed.setDescription(start[srtat] + conf.md + time + " UTC" + tz + conf.md);
+    embed.setTitle("Time.");
+    embed.setTimestamp();
 
     message.reply(embed);
 }